+++ app/components/Form/index.ts
@@ -7,7 +7,7 @@ export { Button } from '@webkom/lego-bricks';
 export { default as Form } from './Form';
 export { default as TimePicker } from './TimePicker';
 export { default as RadioButton } from './RadioButton';
-export { default as RadioButtonGroup } from './RadioButtonGroup';
+export { default as MultiSelectGroup } from './MultiSelectGroup';
 export { default as CheckBox } from './CheckBox';
 export { default as SelectInput } from './SelectInput';
 export { default as PhoneNumberInput } from './PhoneNumberInput';
