+++ app/routes/companyInterest/components/CompanyInterestPage.tsx
@@ -1,6 +1,7 @@
 import { LoadingIndicator } from '@webkom/lego-bricks';
 import arrayMutators from 'final-form-arrays';
-import { Field } from 'react-final-form';
+import { type ReactNode } from 'react';
+import { Field, FormSpy } from 'react-final-form';
 import { FieldArray } from 'react-final-form-arrays';
 import { Helmet } from 'react-helmet-async';
 import { Link } from 'react-router-dom';
@@ -17,7 +18,7 @@ import {
   CheckBox,
   SelectInput,
   RadioButton,
-  RadioButtonGroup,
+  MultiSelectGroup,
 } from 'app/components/Form';
 import LegoFinalForm from 'app/components/Form/LegoFinalForm';
 import Icon from 'app/components/Icon';
@@ -25,100 +26,40 @@ import { Image } from 'app/components/Image';
 import Flex from 'app/components/Layout/Flex';
 import { readmeIfy } from 'app/components/ReadmeLogo';
 import Tooltip from 'app/components/Tooltip';
-import type {
-  CompanyInterestEntity,
-  CompanyInterestCompanyType,
-} from 'app/reducers/companyInterest';
+import type { CompanyInterestEntity } from 'app/reducers/companyInterest';
 import type { CompanySemesterEntity } from 'app/reducers/companySemesters';
 import { spySubmittable, spyValues } from 'app/utils/formSpyUtils';
-import { createValidator, required, isEmail } from 'app/utils/validation';
+import {
+  createValidator,
+  required,
+  isEmail,
+  requiredIf,
+} from 'app/utils/validation';
 import { interestText, semesterToText } from '../utils';
 import styles from './CompanyInterest.css';
-import type { ReactNode } from 'react';
+import {
+  COLLABORATION_TYPES,
+  EVENTS,
+  README,
+  SURVEY_OFFERS,
+  TARGET_GRADES,
+  FORM_LABELS,
+  COMPANY_TYPES,
+  OFFICE_IN_TRONDHEIM,
+} from './Translations';
 
-export const EVENT_TYPES = {
-  company_presentation: {
-    norwegian: 'Bedriftspresentasjon',
-    english: 'Company presentation',
-  },
-  lunch_presentation: {
-    norwegian: 'Lunsjpresentasjon',
-    english: 'Lunch presentation',
-  },
-  course: {
-    norwegian: 'Faglig arrangement',
-    english: 'Course or workshop',
-  },
-  breakfast_talk: {
-    norwegian: 'Frokostforedrag',
-    english: 'Breakfast talk',
-  },
-  // digital_presentation: {
-  //   norwegian: 'Digital presentasjon',
-  //   english: 'Digital presentation',
-  // },
-  bedex: {
-    norwegian: 'Bedriftsekskursjon (BedEx)',
-    english: 'Company excursion (BedEx)',
-  },
-  other: {
-    norwegian: 'Alternativt arrangement',
-    english: 'Other event',
-  },
-  start_up: {
-    norwegian: 'Start-up kveld',
-    english: 'Start-up night',
-  },
-  company_to_company: {
-    norwegian: 'Bedrift-til-bedrift',
-    english: 'Company-to-company',
-  },
+export const PARTICIPANT_RANGE_TYPES = {
+  first: '10-30',
+  second: '30-60',
+  third: '60-100',
+  fourth: '100+',
 };
-export const SURVEY_OFFER_TYPES = {
-  company_survey_security: {
-    norwegian: 'Sikkerhet',
-    english: 'Security',
-  },
-  company_survey_ai: {
-    norwegian: 'Kunstig intelligens',
-    english: 'Artificial intelligence',
-  },
-  company_survey_big_data: {
-    norwegian: 'Big data',
-    english: 'Big data',
-  },
-  company_survey_front_back_end: {
-    norwegian: 'Front- og back-end',
-    english: 'Front- and back-end',
-  },
-  company_survey_iot: {
-    norwegian: 'Internet of Things',
-    english: 'Internet of Things',
-  },
-  company_survey_gamedev: {
-    norwegian: 'Spillutvikling',
-    english: 'Game development',
-  },
-  company_survey_softskills: {
-    norwegian: 'Softskills',
-    english: 'Soft skills',
-  },
-  company_survey_fintech: {
-    norwegian: 'Finansiell teknologi',
-    english: 'Financial technology',
-  },
-};
-export const OTHER_TYPES = {
-  readme: {
-    norwegian: 'Annonse i readme',
-    english: 'Advertisement in readme',
-  },
-  /*
-  collaboration: {
-    norwegian: 'Samarbeid med andre linjeforeninger',
-    english: 'Collaboration with other student organizations',
-  },
-  */
+
+export const PARTICIPANT_RANGE_MAP = {
+  first: [10, 40],
+  second: [30, 60],
+  third: [60, 100],
+  fourth: [100, null],
 };
 
 export const EVENT_TYPE_OPTIONS = [
@@ -135,110 +76,20 @@ export const EVENT_TYPE_OPTIONS = [
   { value: 'company_to_company', label: 'Bedrift-til-bedrift' },
 ];
 
-export const OFFICE_IN_TRONDHEIM = {
-  yes: { norwegian: 'Ja', english: 'Yes' },
-  no: { norwegian: 'Nei', english: 'No' },
-};
-
-export const COLLABORATION_TYPES = {
-  collaboration_omega: {
-    norwegian: 'Samarbeid med Omega linjeforening',
-    english: 'Event in collaboration with Omega',
-  },
-  collaboration_online: {
-    norwegian: 'Samarbeid med Online linjeforening',
-    english: 'Event in collaboration with Online',
-  },
-  collaboration_tihlde: {
-    norwegian: 'Samarbeid med TIHLDE linjeforening',
-    english: 'Event in collaboration with TIHLDE',
-  },
-
-  /*
-  collaboration_anniversary: {
-    english: "Collaboration with Abakus' anniversary committee*",
-    norwegian: 'Samarbeid med Abakus sitt Jubileum*',
-  },
-  collaboration_revue_anniversary: {
-    english: "Collaboration with the revue's anniversary committee*",
-    norwegian: 'Samarbeid med Revyen sitt Jubileum*',
-  },
-  */
-  /*   collaboration_revue: {
-    norwegian: 'Samarbeid med Revyen**',
-    english: 'Collaboration with the revue**',
-  }, */
-};
-export const TARGET_GRADE_TYPES = {
-  '1': {
-    norwegian: '1. klasse',
-    english: '1st Years',
-  },
-  '2': {
-    norwegian: '2. klasse',
-    english: '2nd Years',
-  },
-  '3': {
-    norwegian: '3. klasse',
-    english: '3rd Years',
-  },
-  '4': {
-    norwegian: '4. klasse',
-    english: '4th Years',
-  },
-  '5': {
-    norwegian: '5. klasse',
-    english: '5th Years',
-  },
-};
-export const PARTICIPANT_RANGE_TYPES = {
-  first: '10-30',
-  second: '30-60',
-  third: '60-100',
-  fourth: '100+',
-};
-export const COMPANY_TYPES: Record<
-  CompanyInterestCompanyType,
-  { norwegian: string; english: string }
-> = {
-  company_types_small_consultant: {
-    norwegian: 'Liten konsulentbedrift ( < ~50)',
-    english: 'Small consultant company ( < ~50)',
-  },
-  company_types_medium_consultant: {
-    norwegian: 'Medium konsulentbedrift ( < 400)',
-    english: 'Medium consultant company ( < 400)',
-  },
-  company_types_large_consultant: {
-    norwegian: 'Stor konsulentbedrift ( > 400)',
-    english: 'Large consultant company ( > 400)',
-  },
-  company_types_inhouse: { norwegian: 'In-house', english: 'In-house' },
-  company_types_others: { norwegian: 'Annet', english: 'Other' },
-  company_types_start_up: { norwegian: 'Start-up', english: 'Start-up' },
-  company_types_governmental: { norwegian: 'Statlig', english: 'Governmental' },
-};
-export const PARTICIPANT_RANGE_MAP = {
-  first: [10, 40],
-  second: [30, 60],
-  third: [60, 100],
-  fourth: [100, null],
-};
-
 const eventToString = (event) =>
-  Object.keys(EVENT_TYPES)[Number(event.charAt(event.length - 2))];
+  Object.keys(EVENTS)[Number(event.charAt(event.length - 2))];
 
 const surveyOffersToString = (offer) =>
-  Object.keys(SURVEY_OFFER_TYPES)[Number(offer.charAt(offer.length - 2))];
+  Object.keys(SURVEY_OFFERS)[Number(offer.charAt(offer.length - 2))];
 
 const otherOffersToString = (offer) =>
-  Object.keys(OTHER_TYPES)[Number(offer.charAt(offer.length - 2))];
+  Object.keys(README)[Number(offer.charAt(offer.length - 2))];
 
 const collaborationToString = (collab) =>
   Object.keys(COLLABORATION_TYPES)[Number(collab.charAt(collab.length - 2))];
 
 const targetGradeToString = (targetGrade) =>
-  Object.keys(TARGET_GRADE_TYPES)[
+  Object.keys(TARGET_GRADES)[
     Number(targetGrade.charAt(targetGrade.length - 2))
   ];
 
@@ -275,7 +126,7 @@ const SurveyOffersBox = ({
       <Field
         key={`companyCourseThemes[${index}]`}
         name={`companyCourseThemes[${index}].checked`}
-        label={SURVEY_OFFER_TYPES[surveyOffersToString(item)][language]}
+        label={SURVEY_OFFERS[surveyOffersToString(item)][language]}
         type="checkbox"
         component={CheckBox.Field}
         normalize={(v) => !!v}
@@ -291,18 +142,32 @@ const EventBox = ({
   fields: any;
   language: string;
 }): ReactNode => (
-  <Flex column className={styles.checkboxWrapper}>
-    {fields.map((key, index) => (
-      <Field
-        key={`events[${index}]`}
-        name={`events[${index}].checked`}
-        label={EVENT_TYPES[eventToString(key)][language]}
-        type="checkbox"
-        component={CheckBox.Field}
-        normalize={(v) => !!v}
-      />
-    ))}
-  </Flex>
+  <FormSpy subscription={{ values: true }}>
+    {(props) => {
+      const filteredFields = fields.map((field) => field); // This is just to get an array instead of what fields is (which is an object that mimics an iterable). See: https://github.com/final-form/react-final-form-arrays#fieldarrayrenderprops
+      if (props.values.officeInTrondheim !== 'yes') {
+        fields.forEach((field, index) => {
+          if (fields.value[index].name == 'company_to_company') {
+            filteredFields.splice(index, 1);
+          }
+        });
+      }
+      return (
+        <Flex column className={styles.checkboxWrapper}>
+          {filteredFields.map((key, index) => (
+            <Field
+              key={`events[${index}]`}
+              name={`events[${index}].checked`}
+              label={EVENTS[eventToString(key)][language]}
+              type="checkbox"
+              component={CheckBox.Field}
+              normalize={(v) => !!v}
+            />
+          ))}
+        </Flex>
+      );
+    }}
+  </FormSpy>
 );
 
 const TargetGradeBox = ({
@@ -317,7 +182,7 @@ const TargetGradeBox = ({
       <Field
         key={`targetGrades[${index}]`}
         name={`targetGrades[${index}].checked`}
-        label={TARGET_GRADE_TYPES[targetGradeToString(key)][language]}
+        label={TARGET_GRADES[targetGradeToString(key)][language]}
         type="checkbox"
         component={CheckBox.Field}
         normalize={(v) => !!v}
@@ -338,7 +203,7 @@ const OtherBox = ({
       <Field
         key={`otherOffers[${index}]`}
         name={`otherOffers[${index}].checked`}
-        label={readmeIfy(OTHER_TYPES[otherOffersToString(key)][language])}
+        label={readmeIfy(README[otherOffersToString(key)][language])}
         type="checkbox"
         component={CheckBox.Field}
         normalize={(v) => !!v}
@@ -449,12 +314,31 @@ type Props = {
   initialValues: any;
 };
 
+const requiredIfEventType = (eventType: string) =>
+  requiredIf((allValues) => {
+    const event = allValues.events.filter(
+      (event) => event.name === eventType
+    )[0];
+    return event && event.checked;
+  });
+
 const validate = createValidator({
   company: [required()],
   contactPerson: [required()],
   mail: [required(), isEmail()],
   phone: [required()],
   comment: [required()],
+  companyType: [required()],
+  officeInTrondheim: [required()],
+  events: [required()],
+  breakfastTalkComment: [requiredIfEventType('breakfast_talk')],
+  companyPresentationComment: [requiredIfEventType('company_presentation')],
+  lunchPresentationComment: [requiredIfEventType('lunsh_presentation')],
+  courseComment: [requiredIfEventType('course')],
+  bedexComment: [requiredIfEventType('bedex')],
+  otherEventComment: [requiredIfEventType('other')],
+  startupComment: [requiredIfEventType('start_up')],
+  companyToCompanyComment: [requiredIfEventType('company_to_company')],
 });
 
 const CompanyInterestPage = (props: Props) => {
@@ -523,104 +407,6 @@ const CompanyInterestPage = (props: Props) => {
       });
   };
 
-  const labels = {
-    mainHeading: {
-      norwegian: 'Meld interesse',
-      english: 'Register interest',
-    },
-    subHeading: {
-      norwegian:
-        'Dette skjemaet skal ikke brukes for annonser. For slikt, send en e-post til ',
-      english:
-        'This form is not to be used for job listings. For such enquiries, send an e-mail to ',
-    },
-    company: {
-      header: {
-        norwegian: 'Navn på bedrift',
-        english: 'Name of company',
-      },
-      placeholder: {
-        norwegian: 'Bedriftsnavn',
-        english: 'Company name',
-      },
-    },
-    officeInTrondheim: {
-      norwegian: 'Har dere kontorer i Trondheim egnet for besøk?',
-      english: 'Do you have offices in Trondheim suited for visiting?',
-    },
-    contactPerson: {
-      header: {
-        norwegian: 'Kontaktperson',
-        english: 'Contact person',
-      },
-      placeholder: {
-        norwegian: 'Kari Nordmann',
-        english: 'Jon Smith',
-      },
-    },
-    mail: {
-      norwegian: 'E-post',
-      english: 'E-mail',
-    },
-    phone: {
-      norwegian: 'Telefonnummer',
-      english: 'Phone number',
-    },
-    semester: {
-      norwegian: 'Semester',
-      english: 'Semester',
-    },
-    events: {
-      norwegian: 'Arrangementer',
-      english: 'Events',
-    },
-    otherOffers: {
-      norwegian: 'Annet',
-      english: 'Other',
-    },
-    companyTypes: {
-      norwegian: 'Bedriftstype',
-      english: 'Company type',
-    },
-    collaborations: {
-      norwegian: 'Samarbeid',
-      english: 'Collaboration',
-    },
-    targetGrades: {
-      norwegian: 'Klassetrinn',
-      english: 'Target grades',
-    },
-    companyCourseThemes: {
-      norwegian: 'Temaer som er relevant for dere',
-      english: 'Topics that are relevant for you',
-    },
-    participantRange: {
-      norwegian: 'Antall deltagere',
-      english: 'Number of participants',
-    },
-    comment: {
-      norwegian: 'Om bedriften',
-      english: 'About the company',
-    },
-    secondComment: {
-      norwegian: 'Annen kommentar',
-      english: 'Other comment',
-    },
-    create: {
-      norwegian: 'Send bedriftsinteresse',
-      english: 'Submit interest',
-    },
-    eventDescriptionHeader: {
-      norwegian: 'Pitch/forklar dine ønsker for arrangementet',
-      english: 'Pitch/explain your wishes for the event',
-    },
-    eventDescriptionIntro: {
-      norwegian:
-        'Skriv gjerne litt om hvilke type arrangementer dere ønsker å arrangere. Vi prøver å planlegge med flere ulike typer arrangementer og bedrifter der vi prøver å lage et variert, spennende og nyskapende program. Våre bedriftskontakter har også muligheten til å hjelpe med å utvikle gode arrangementer.',
-      english:
-        'Please write a bit about what types of events you would like to arrange. We try to plan with several different types of events and companies, where we try to create a varied, exciting and innovative program. Our company contacts also have the opportunity to help develop good events.',
-    },
-  };
   const { language } = props;
   const isEnglish = language === 'english';
 
@@ -640,7 +426,7 @@ const CompanyInterestPage = (props: Props) => {
         {({ handleSubmit }) => (
           <form onSubmit={handleSubmit}>
             <FlexRow alignItems="center" justifyContent="space-between">
-              <h1>{labels.mainHeading[language]}</h1>
+              <h1>{FORM_LABELS.mainHeading[language]}</h1>
               {!props.edit && (
                 <Link to={isEnglish ? '/interesse' : '/register-interest'}>
                   <LanguageFlag language={language} />
@@ -648,7 +434,7 @@ const CompanyInterestPage = (props: Props) => {
               )}
             </FlexRow>
             <Card severity="info">
-              {labels.subHeading[language]}
+              {FORM_LABELS.subHeading[language]}
               <a href={'mailto:bedriftskontakt@abakus.no'}>
                 bedriftskontakt@abakus.no
               </a>
@@ -656,29 +442,29 @@ const CompanyInterestPage = (props: Props) => {
 
             <Field
               name="company"
-              label={labels.company.header[language]}
-              placeholder={labels.company.placeholder[language]}
+              label={FORM_LABELS.company.header[language]}
+              placeholder={FORM_LABELS.company.placeholder[language]}
               filter={['companies.company']}
               fieldClassName={styles.metaField}
               component={SelectInput.AutocompleteField}
               required
             />
             <Field
-              label={labels.contactPerson.header[language]}
-              placeholder={labels.contactPerson.placeholder[language]}
+              label={FORM_LABELS.contactPerson.header[language]}
+              placeholder={FORM_LABELS.contactPerson.placeholder[language]}
               name="contactPerson"
               component={TextInput.Field}
               required
             />
             <Field
-              label={labels.mail[language]}
+              label={FORM_LABELS.mail[language]}
               placeholder="example@gmail.com"
               name="mail"
               component={TextInput.Field}
               required
             />
             <Field
-              label={labels.phone[language]}
+              label={FORM_LABELS.phone[language]}
               placeholder="+47 909 09 090"
               name="phone"
               component={TextInput.Field}
@@ -688,9 +474,9 @@ const CompanyInterestPage = (props: Props) => {
             <Flex wrap justifyContent="space-between">
               <Flex column className={styles.interestBox}>
                 <label htmlFor="companyType" className={styles.heading}>
-                  {labels.companyTypes[language]}
+                  {FORM_LABELS.companyTypes[language]}
                 </label>
-                <RadioButtonGroup name="companyType">
+                <MultiSelectGroup name="companyType">
                   {Object.keys(COMPANY_TYPES).map((key) => (
                     <Field
                       key={key}
@@ -699,15 +485,16 @@ const CompanyInterestPage = (props: Props) => {
                       type="radio"
                       component={RadioButton.Field}
                       value={key}
+                      showErrors={false}
                     />
                   ))}
-                </RadioButtonGroup>
+                </MultiSelectGroup>
               </Flex>
               <Flex column className={styles.interestBox}>
                 <label htmlFor="officeInTrondheim" className={styles.heading}>
-                  {labels.officeInTrondheim[language]}
+                  {FORM_LABELS.officeInTrondheim[language]}
                 </label>
-                <RadioButtonGroup name="officeInTrondheim">
+                <MultiSelectGroup name="officeInTrondheim">
                   {Object.keys(OFFICE_IN_TRONDHEIM).map((key) => (
                     <Field
                       key={key}
@@ -716,21 +503,22 @@ const CompanyInterestPage = (props: Props) => {
                       type="radio"
                       component={RadioButton.Field}
                       value={key}
+                      showErrors={false}
                     />
                   ))}
-                </RadioButtonGroup>
+                </MultiSelectGroup>
               </Flex>
               <Flex column className={styles.interestBox}>
                 <label htmlFor="companyCourseThemes" className={styles.heading}>
                   <Flex alignItems="center" gap="5px">
-                    {labels.companyCourseThemes[language]}
+                    {FORM_LABELS.companyCourseThemes[language]}
                     <Tooltip
                       className={styles.tooltip}
                       content={
                         <span>
                           {language === 'norwegian'
-                            ? 'Dette er temaer som studenter uttrykte interesse for å lære mer om i vår bedriftsundersøkelse. Er dette relevante temaer for deres bedrift dere kan være interessert i å avholde faglige arrangementer eller snakke om på deres presentasjoner? (Uforpliktende)'
-                            : 'These are topics that students expressed interest in learning more about in our company survey. Are these relevant topics for your company that you might be interested in arranging a course or workshop about or talk about in your presentations? (Non-binding'}
+                            ? 'Dette er temaer som studenter har uttrykt interesse for å lære mer om i vår bedriftsundersøkelse. Kryss av for de temaene dere kan ønske å holde kurs om eller snakke om på deres presentasjoner. (Uforpliktende)'
+                            : 'These are topics that students expressed interest in learning more about in our company survey. Check off the topics that you might be interested in arranging a course or workshop about or talk about in your presentations. (Non-binding)'}
                         </span>
                       }
                     >
@@ -738,7 +526,6 @@ const CompanyInterestPage = (props: Props) => {
                     </Tooltip>
                   </Flex>
                 </label>
-
                 <FieldArray
                   label="companyCourseThemes"
                   name="companyCourseThemes"
@@ -747,39 +534,20 @@ const CompanyInterestPage = (props: Props) => {
                 />
               </Flex>
             </Flex>
-            <div className={styles.topline}>
-              {interestText.text.first[language]}
-              {/*
-          <br />
-          <br />
-          {interestText.text.second[language]}
-          <br />
-          <br />
-          {interestText.bedex[language]
-          <br />
-          <br />
-          {interestText.anniversaryCollaboration[language]}
-          */}
-              {/*           <br />
-          <br />
-          {interestText.revueCollaboration[language]}
- */}{' '}
-            </div>
-
+            <div className={styles.topline} />
             <Field
               placeholder={interestText.comment[language]}
               name="comment"
               component={TextEditor.Field}
               rows={10}
               className={styles.textEditor}
-              label={labels.comment[language]}
+              label={FORM_LABELS.comment[language]}
               required
             />
-
             <Flex wrap justifyContent="space-between">
               <Flex column className={styles.interestBox}>
                 <label htmlFor="semesters" className={styles.heading}>
-                  {labels.semester[language]}
+                  {FORM_LABELS.semester[language]}
                 </label>
                 <FieldArray
                   label="semesters"
@@ -788,19 +556,21 @@ const CompanyInterestPage = (props: Props) => {
                   component={SemesterBox}
                 />
               </Flex>
-              <Flex column className={styles.interestBox}>
-                <label htmlFor="events" className={styles.heading}>
-                  {labels.events[language]}
-                </label>
-                <FieldArray
-                  name="events"
-                  language={language}
-                  component={EventBox}
-                />
-              </Flex>
+              <MultiSelectGroup name="events">
+                <Flex column className={styles.interestBox}>
+                  <label htmlFor="events" className={styles.heading}>
+                    {FORM_LABELS.events[language]}
+                  </label>
+                  <FieldArray
+                    name="events"
+                    language={language}
+                    component={EventBox}
+                  />
+                </Flex>
+              </MultiSelectGroup>
               <Flex column className={styles.interestBox}>
                 <label htmlFor="collaborations" className={styles.heading}>
-                  {labels.collaborations[language]}
+                  {FORM_LABELS.collaborations[language]}
                 </label>
                 <FieldArray
                   name="collaborations"
@@ -813,7 +583,7 @@ const CompanyInterestPage = (props: Props) => {
             <Flex wrap justifyContent="space-between">
               <Flex column className={styles.interestBox}>
                 <label htmlFor="targetGrades" className={styles.heading}>
-                  {labels.targetGrades[language]}
+                  {FORM_LABELS.targetGrades[language]}
                 </label>
                 <FieldArray
                   name="targetGrades"
@@ -824,9 +594,9 @@ const CompanyInterestPage = (props: Props) => {
 
               <Flex column className={styles.interestBox}>
                 <label htmlFor="participantRange" className={styles.heading}>
-                  {labels.participantRange[language]}
+                  {FORM_LABELS.participantRange[language]}
                 </label>
-                <RadioButtonGroup name="participantRange">
+                <MultiSelectGroup name="participantRange">
                   {Object.keys(PARTICIPANT_RANGE_TYPES).map((key) => (
                     <Field
                       key={key}
@@ -837,11 +607,11 @@ const CompanyInterestPage = (props: Props) => {
                       value={key}
                     />
                   ))}
-                </RadioButtonGroup>
+                </MultiSelectGroup>
               </Flex>
               <Flex column className={styles.interestBox}>
                 <label htmlFor="otherOffers" className={styles.heading}>
-                  {labels.otherOffers[language]}
+                  {FORM_LABELS.otherOffers[language]}
                 </label>
                 <FieldArray
                   name="otherOffers"
@@ -851,9 +621,9 @@ const CompanyInterestPage = (props: Props) => {
               </Flex>
             </Flex>
             <h3 className={styles.topline}>
-              {labels.eventDescriptionHeader[language]}
+              {FORM_LABELS.eventDescriptionHeader[language]}
             </h3>
-            <p>{labels.eventDescriptionIntro[language]}</p>
+            <p>{FORM_LABELS.eventDescriptionIntro[language]}</p>
 
             {spyValues((values: CompanyInterestFormEntity) => {
               const showCompanyPresentation = values.events?.some(
@@ -864,7 +634,7 @@ const CompanyInterestPage = (props: Props) => {
                 showCompanyPresentation && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.company_presentation[language]}</h4>
+                      <h4>{EVENTS.company_presentation[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>
@@ -893,7 +663,7 @@ const CompanyInterestPage = (props: Props) => {
                 showLunchPresentationComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.lunch_presentation[language]}</h4>
+                      <h4>{EVENTS.lunch_presentation[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>
@@ -922,7 +692,7 @@ const CompanyInterestPage = (props: Props) => {
                 showCourseComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.course[language]}</h4>
+                      <h4>{EVENTS.course[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>{interestText.courseDescription[language]}</p>
@@ -947,7 +717,7 @@ const CompanyInterestPage = (props: Props) => {
                 showBreakfastTalkComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.breakfast_talk[language]}</h4>
+                      <h4>{EVENTS.breakfast_talk[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>{interestText.breakfastTalkDescription[language]}</p>
@@ -972,7 +742,7 @@ const CompanyInterestPage = (props: Props) => {
                 showBedexComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.bedex[language]}</h4>
+                      <h4>{EVENTS.bedex[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>{interestText.bedexDescription[language]}</p>
@@ -997,7 +767,7 @@ const CompanyInterestPage = (props: Props) => {
                 showOtherEventComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.other[language]}</h4>
+                      <h4>{EVENTS.other[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>{interestText.otherEventDescription[language]}</p>
@@ -1022,7 +792,7 @@ const CompanyInterestPage = (props: Props) => {
                 showStartupComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.start_up[language]}</h4>
+                      <h4>{EVENTS.start_up[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
                     <p>{interestText.startUpDescription[language]}</p>
@@ -1047,7 +817,7 @@ const CompanyInterestPage = (props: Props) => {
                 showCompanyToCompanyComment && (
                   <div className={styles.topline}>
                     <Flex alignItems="center" gap={1}>
-                      <h4>{EVENT_TYPES.company_to_company[language]}</h4>
+                      <h4>{EVENTS.company_to_company[language]}</h4>
                       <p className={styles.label}>*</p>
                     </Flex>
 
@@ -1076,7 +846,7 @@ const CompanyInterestPage = (props: Props) => {
               <Button secondary disabled={!submittable} submit>
                 {props.edit
                   ? 'Oppdater bedriftsinteresse'
-                  : labels.create[language]}
+                  : FORM_LABELS.create[language]}
               </Button>
             ))}
           </form>
